+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.3.1-SNAPSHOT
 
-
+* [FEATURE] Check last version at start
 
 ## 2.3.0 / 2022-04-09